// client/src/components/Common/LaserFlow/LaserFlow.d.ts
import type { CSSProperties } from 'react';

export interface LaserFlowProps {
  className?: string;
  style?: CSSProperties;
  wispDensity?: number;
  dpr?: number;
  mouseSmoothTime?: number;
  mouseTiltStrength?: number;
  horizontalBeamOffset?: number;
  verticalBeamOffset?: number;
  flowSpeed?: number;
  verticalSizing?: number;
  horizontalSizing?: number;
  fogIntensity?: number;
  fogScale?: number;
  wispSpeed?: number;
  wispIntensity?: number;
  flowStrength?: number;
  decay?: number;
  falloffStart?: number;
  fogFallSpeed?: number;
  color?: string;
}

declare const LaserFlow: React.FC<LaserFlowProps>;
export default LaserFlow;